requirements.txt
envparse==0.2.0
peewee==2.8.5
PyStemmer==1.3.0
python-telegram-bot==5.3.0
